Added support for proper reporting of the number of bytes sent on a write that only partially completes.
git-svn-id: https://svn.code.sf.net/p/linux-gpib/code/trunk@1371 b61deed1-901b-0410-a6dd-e94ecbdaea89

+++ linux-gpib/driver/sys/ibread.c
@@ -32,7 +32,7 @@
  *          calling ibcmd.
  */
 
-ssize_t ibrd(gpib_board_t *board, uint8_t *buf, size_t length, int *end_flag, int *nbytes)
+int ibrd(gpib_board_t *board, uint8_t *buf, size_t length, int *end_flag, size_t *nbytes)
 {
 	ssize_t ret = 0;
 	int retval;
